rename : rename model package to domainEntity
