+++ src/main/java/roomescape/RoomescapeApplication.java
@@ -1,9 +1,13 @@
 package roomescape;
 
 import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication
+//@SpringBootApplication
+@EnableAutoConfiguration
+//@ComponentScan
+//@Configuration
 public class RoomescapeApplication {
     public static void main(String[] args) {
         SpringApplication.run(RoomescapeApplication.class, args);
