+++ src/main/java/roomescape/domain_entity/Reservations.java
@@ -1,4 +1,4 @@
-package roomescape.model;
+package roomescape.domain_entity;
 
 import java.util.ArrayList;
 import java.util.Collections;
