+++ src/main/java/roomescape/dto/ReservationRequestDto.java
@@ -2,9 +2,7 @@ package roomescape.dto;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
-import java.time.format.DateTimeFormatter;
-import roomescape.model.Id;
-import roomescape.model.Reservation;
+import roomescape.domain_entity.Reservation;
 
 public record ReservationRequestDto(String name, LocalDate date, LocalTime time) {
